Convert Egg/Fossil Bots into Encounter Bots
Adds option to reboots the game when fossil pieces are depleted, and
user chooses not to InjectWhenEmpty.

Moves all settings for Egg/Fossil bots under EncounterSWSH.

Egg Bot may be a little faster than before.

Various error messages tweaked for clarity.
