--- SysBot.Pokemon/SWSH/BotEgg/EggSettings.cs
@@ -1,42 +1,0 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel;
-using System.Threading;
-using SysBot.Base;
-
-namespace SysBot.Pokemon
-{
-    public class EggSettings : IBotStateSettings, ICountSettings
-    {
-        private const string FeatureToggle = nameof(FeatureToggle);
-        private const string Counts = nameof(Counts);
-        public override string ToString() => "Egg Bot Settings";
-
-        [Category(FeatureToggle), Description("When enabled, the bot will continue after finding a suitable match.")]
-        public ContinueAfterMatch ContinueAfterMatch { get; set; } = ContinueAfterMatch.StopExit;
-
-        [Category(FeatureToggle), Description("When enabled, the screen will be turned off during normal bot loop operation to save power.")]
-        public bool ScreenOff { get; set; }
-
-        private int _completedEggs;
-
-        [Category(Counts), Description("Eggs Retrieved")]
-        public int CompletedEggs
-        {
-            get => _completedEggs;
-            set => _completedEggs = value;
-        }
-
-        [Category(Counts), Description("When enabled, the counts will be emitted when a status check is requested.")]
-        public bool EmitCountsOnStatusCheck { get; set; }
-
-        public int AddCompletedEggs() => Interlocked.Increment(ref _completedEggs);
-
-        public IEnumerable<string> GetNonZeroCounts()
-        {
-            if (!EmitCountsOnStatusCheck)
-                yield break;
-            if (CompletedEggs != 0)
-                yield return $"Eggs Received: {CompletedEggs}";
-        }
-    }
-}
