+++ SysBot.Pokemon/SWSH/BotEncounter/EncounterBotFossil/FossilSpecies.cs
