+++ SysBot.Pokemon/SWSH/BotEncounter/EncounterSettings.cs
@@ -9,10 +9,15 @@ namespace SysBot.Pokemon
     {
         private const string Counts = nameof(Counts);
         private const string Encounter = nameof(Encounter);
+        private const string Settings = nameof(Settings);
         public override string ToString() => "Encounter Bot SWSH Settings";
 
-        [Category(Encounter), Description("The method by which the Line bot will encounter Pokémon.")]
+        [Category(Encounter), Description("The method used by the Line and Reset bots to encounter Pokémon.")]
         public EncounterMode EncounteringType { get; set; } = EncounterMode.VerticalLine;
+
+        [Category(Settings)]
+        [TypeConverter(typeof(ExpandableObjectConverter))]
+        public FossilSettings Fossil { get; set; } = new();
 
         [Category(Encounter), Description("When enabled, the bot will continue after finding a suitable match.")]
         public ContinueAfterMatch ContinueAfterMatch { get; set; } = ContinueAfterMatch.StopExit;
@@ -22,6 +27,8 @@ namespace SysBot.Pokemon
 
         private int _completedWild;
         private int _completedLegend;
+        private int _completedEggs;
+        private int _completedFossils;
 
         [Category(Counts), Description("Encountered Wild Pokémon")]
         public int CompletedEncounters
@@ -37,11 +44,28 @@ namespace SysBot.Pokemon
             set => _completedLegend = value;
         }
 
+        [Category(Counts), Description("Eggs Retrieved")]
+        public int CompletedEggs
+        {
+            get => _completedEggs;
+            set => _completedEggs = value;
+        }
+
+
+        [Category(Counts), Description("Fossil Pokémon Revived")]
+        public int CompletedFossils
+        {
+            get => _completedFossils;
+            set => _completedFossils = value;
+        }
+
         [Category(Counts), Description("When enabled, the counts will be emitted when a status check is requested.")]
         public bool EmitCountsOnStatusCheck { get; set; }
 
         public int AddCompletedEncounters() => Interlocked.Increment(ref _completedWild);
         public int AddCompletedLegends() => Interlocked.Increment(ref _completedLegend);
+        public int AddCompletedEggs() => Interlocked.Increment(ref _completedEggs);
+        public int AddCompletedFossils() => Interlocked.Increment(ref _completedFossils);
 
         public IEnumerable<string> GetNonZeroCounts()
         {
@@ -51,6 +75,10 @@ namespace SysBot.Pokemon
                 yield return $"Wild Encounters: {CompletedEncounters}";
             if (CompletedLegends != 0)
                 yield return $"Legendary Encounters: {CompletedLegends}";
+            if (CompletedEggs != 0)
+                yield return $"Eggs Received: {CompletedEggs}";
+            if (CompletedFossils != 0)
+                yield return $"Completed Fossils: {CompletedFossils}";
         }
     }
 }
