--- SysBot.Pokemon/SWSH/BotFossil/FossilBot.cs
@@ -1,183 +1,0 @@
-﻿using PKHeX.Core;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using SysBot.Base;
-using static SysBot.Base.SwitchButton;
-using static SysBot.Pokemon.PokeDataOffsets;
-
-namespace SysBot.Pokemon
-{
-    public class FossilBot : PokeRoutineExecutor8, IEncounterBot
-    {
-        private readonly PokeTradeHub<PK8> Hub;
-        private readonly FossilSettings Settings;
-        private readonly IDumper DumpSetting;
-        private readonly int[] DesiredMinIVs;
-        private readonly int[] DesiredMaxIVs;
-        public ICountSettings Counts => Settings;
-
-        public FossilBot(PokeBotState cfg, PokeTradeHub<PK8> hub) : base(cfg)
-        {
-            Hub = hub;
-            Settings = Hub.Config.FossilSWSH;
-            DumpSetting = Hub.Config.Folder;
-            StopConditionSettings.InitializeTargetIVs(Hub.Config, out DesiredMinIVs, out DesiredMaxIVs);
-        }
-
-        // Cached offsets that stay the same per session.
-        private ulong OverworldOffset;
-
-        private int encounterCount;
-
-        private static readonly PK8 Blank = new();
-
-        public override async Task MainLoop(CancellationToken token)
-        {
-            Log("Identifying trainer data of the host console.");
-            await IdentifyTrainer(token).ConfigureAwait(false);
-            await SetupBoxState(DumpSetting, token).ConfigureAwait(false);
-
-            OverworldOffset = await SwitchConnection.PointerAll(Offsets.OverworldPointer, token).ConfigureAwait(false);
-
-            Log("Checking item counts...");
-            var pouchData = await Connection.ReadBytesAsync(ItemTreasureAddress, 80, token).ConfigureAwait(false);
-            var counts = FossilCount.GetFossilCounts(pouchData);
-            int reviveCount = counts.PossibleRevives(Settings.Species);
-            if (reviveCount == 0)
-            {
-                Log("Insufficient fossil pieces. Please obtain at least one of each required fossil piece first.");
-                return;
-            }
-
-            try
-            {
-                await InitializeHardware(Settings, token).ConfigureAwait(false);
-                Log($"Starting main {nameof(FossilBot)} loop.");
-                Config.IterateNextRoutine();
-                await InnerLoop(reviveCount, pouchData, counts, token).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                Log(e.Message);
-            }
-
-            Log($"Ending {nameof(FossilBot)} loop.");
-            await HardStop().ConfigureAwait(false);
-        }
-
-        private async Task InnerLoop(int reviveCount, byte[] pouchData, FossilCount counts, CancellationToken token)
-        {
-            while (!token.IsCancellationRequested && Config.NextRoutineType == PokeRoutineType.FossilBot)
-            {
-                if (encounterCount != 0 && encounterCount % reviveCount == 0)
-                {
-                    Log($"Ran out of fossils to revive {Settings.Species}.");
-                    if (Settings.InjectWhenEmpty)
-                    {
-                        Log("Restoring original pouch data.");
-                        await Connection.WriteBytesAsync(pouchData, ItemTreasureAddress, token).ConfigureAwait(false);
-                        await Task.Delay(500, token).ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        Log("Restart the game and the bot(s) or set \"Inject Fossils\" to True in the config.");
-                        return;
-                    }
-                }
-
-                Log("Clearing destination slot.");
-                await SetBoxPokemon(Blank, 0, 0, token).ConfigureAwait(false);
-                await ReviveFossil(counts, token).ConfigureAwait(false);
-                Log("Fossil revived. Checking details...");
-
-                var pk = await ReadBoxPokemon(0, 0, token).ConfigureAwait(false);
-                if (pk.Species == 0 || !pk.ChecksumValid)
-                {
-                    Log("Invalid data detected in destination slot. Restarting loop.");
-                    continue;
-                }
-
-                encounterCount++;
-                var print = Hub.Config.StopConditions.GetPrintName(pk);
-                Log($"Encounter: {encounterCount}{Environment.NewLine}{print}{Environment.NewLine}");
-                if (DumpSetting.Dump)
-                    DumpPokemon(DumpSetting.DumpFolder, "fossil", pk);
-
-                Settings.AddCompletedFossils();
-
-                if (!StopConditionSettings.EncounterFound(pk, DesiredMinIVs, DesiredMaxIVs, Hub.Config.StopConditions, null))
-                    continue;
-
-                if (Hub.Config.StopConditions.CaptureVideoClip)
-                {
-                    await Task.Delay(Hub.Config.StopConditions.ExtraTimeWaitCaptureVideo, token).ConfigureAwait(false);
-                    await PressAndHold(CAPTURE, 2_000, 0, token).ConfigureAwait(false);
-                }
-
-                var mode = Settings.ContinueAfterMatch;
-                var msg = $"Result found!\n{print}\n" + mode switch
-                {
-                    ContinueAfterMatch.Continue             => "Continuing...",
-                    ContinueAfterMatch.PauseWaitAcknowledge => "Waiting for instructions to continue.",
-                    ContinueAfterMatch.StopExit             => "Stopping routine execution; restart the bot to search again.",
-                    _ => throw new ArgumentOutOfRangeException(),
-                };
-
-                if (!string.IsNullOrWhiteSpace(Hub.Config.StopConditions.MatchFoundEchoMention))
-                    msg = $"{Hub.Config.StopConditions.MatchFoundEchoMention} {msg}";
-                EchoUtil.Echo(msg);
-                Log(msg);
-
-                if (mode == ContinueAfterMatch.StopExit)
-                    break;
-                if (mode == ContinueAfterMatch.Continue)
-                    continue;
-
-                IsWaiting = true;
-                while (IsWaiting)
-                    await Task.Delay(1_000, token).ConfigureAwait(false);
-            }
-        }
-
-        public override async Task HardStop()
-        {
-            await CleanExit(CancellationToken.None).ConfigureAwait(false);
-        }
-
-        private bool IsWaiting;
-        public void Acknowledge() => IsWaiting = false;
-
-        private async Task ReviveFossil(FossilCount count, CancellationToken token)
-        {
-            Log("Starting fossil revival routine...");
-            if (GameLang == LanguageID.Spanish)
-                await Click(A, 0_900, token).ConfigureAwait(false);
-
-            await Click(A, 1_100, token).ConfigureAwait(false);
-
-            // French is slightly slower.
-            if (GameLang == LanguageID.French)
-                await Task.Delay(0_200, token).ConfigureAwait(false);
-
-            await Click(A, 1_300, token).ConfigureAwait(false);
-
-            // Selecting first fossil.
-            if (count.UseSecondOption1(Settings.Species))
-                await Click(DDOWN, 0_300, token).ConfigureAwait(false);
-            await Click(A, 1_300, token).ConfigureAwait(false);
-
-            // Selecting second fossil.
-            if (count.UseSecondOption2(Settings.Species))
-                await Click(DDOWN, 300, token).ConfigureAwait(false);
-
-            // A spam through accepting the fossil and agreeing to revive.
-            for (int i = 0; i < 8; i++)
-                await Click(A, 0_400, token).ConfigureAwait(false);
-
-            // Safe to mash B from here until we get out of all menus.
-            while (!await IsOnOverworld(OverworldOffset, token).ConfigureAwait(false))
-                await Click(B, 0_400, token).ConfigureAwait(false);
-        }
-    }
-}
